Fight order: stop gold medal fights being placed before other fights in the order. Golden Score: allow seperation based on time and weight Fight Points: change simple score from int to double

+++ src/au/com/jwatmuff/eventmanager/gui/results/ResultInfoMapper.java
@@ -42,7 +42,7 @@ public class ResultInfoMapper implements BeanMapper<ResultInfo> {
             map.put("playerId" + (i + 1), (player != null) ? player.getVisibleID() : "N/A");
         }
 
-        int[] scores = bean.getResult().getSimpleScores(database);
+        double[] scores = bean.getResult().getSimpleScores(database);
         map.put("score", scores[0] + " : " + scores[1]);
         if(scores[0] > scores[1])
             map.put("winner", bean.getPlayerName()[0]);
